Initial commit. This is a work in progress of a REST extension for MapGuide built on top of the Slim PHP micro-framework.
This REST extension is functionally similar to mapguide4j, except it is PHP-based and has a much more simplified install story (just drop this dir into the www directory of MapGuide and away you go!). Slim was chosen for its micro-framework nature (we just wanted a powerful, flexible router) with good performance.

The current aim of this project is:
 - To replicate the existing REST API of mapguide4j
 - To provide a GeoREST-style feature of configuring specific Feature Sources to be accessible in a REST-ful fashion with various representations.