Add template REST adapter. This gives us a flexible template-based representation whose actual content can be customized by the user.
We use Smarty for our templating needs. Smarty was chosen because it supported while loop constructs in its templates, which we require because our main input source is a MgReader. From a template perspective, they are passed a MgFeatureModel for single-result templates, a MgFeatureReaderModel for many-result templates, a MgException for error templates and the identity property value for none templates.

An example of these HTML templates is included under the property configuration.

Also in this submission, all references to Pimple have been removed. Slim already provides a DI container, making Pimple redundant.