Add KML support. This includes the following routes:
 * /library/{resourcePath}.MapDefinition/kml
 * /library/{resourcePath}.LayerDefinition/kml
 * /library/{resourcePath}.LayerDefinition/kmlfeatures

All of these routes wrap the existing KML service APIs or duplicate the native MapGuide implementation to generate the NetworkLink elements.

Also add support for chunking response content from a MgHttpResponse if a X-CHUNK-RESPONSE parameter is specified. The /library/{resourcePath}.LayerDefinition/kmlfeatures route will output via MgHttpChunkWriter

For performance, the /library/{resourcePath}.MapDefinition/kml route also offers a pass-through option (which is enabled by default). In pass-through mode, the request is redirected to the equivalent mapagent URL and this is reflected in all the NetworkLink elements that are generated. When not passing through, these NetworkLinks will point to their downstream RESTful URL counterparts.

Fixes #74