#105: When a "many" data query yields no results, re-route to the "none" template view. This adds a new variable to the "none" template: single
single is a boolean which if true, means that the ID variable will also be defined. If single is false, the "none" view was shown from a "many" data query. This allows the "none" view to have custom content based on whether the failing query was a "single" or "many" query.