Add peeking support (via Peek()) to MgFeatureReaderModel. A call to Peek() is the same as a call to Next() except that a call to Peek() will cache the peeked record as the current record and calling Next() after a Peek() does not advance the internal feature reader, but rather resets the current record to the peeked one.
The purpose of this peeking support is to allow us to determine whether a feature reader has a result (by peeking ahead) without advancing the reader (from the perspective of the calling side). This will allow us in the template adapter to then have a multiple feature query result with one record be re-routed to the "single" template view instead of the "many" template view.

New tests included to test peeking