This mega-commit does some preparatory refactoring to make the upgrade from slim 2 to slim 3 much easier.
Slim 3 decouples request and response from the main slim app, so this refactoring is to abstract away the many accesses to $this->app, $this->app->request and $this->app->response behind helper methods in the base MgResponseHandler class. Once we are ready to migrate to slim 3, we can then just update these helper methods and fix up the constructor parameters being passed down.